Recovery combinators not working, type annotations needed...
The I of Recovery cannot be deduced. It must be the same as the I
of Parser, but there's no way that I can see to specify that without
equality constraints or turning I into an associated type. The latter
can't be done because then Parser::Error needs to depend on
Parser::Input, which is not object-safe because reasons. Turning I
of Recovery into Recovery::Input turned into a disaster almost
instantly so that's not it either.

Maybe it can be resolved by making ToRecover the Nothing strategy,
and blanket-implementing Parser for all strategies using a
parse_internal() that does the same thing as ToRecover does now,
but which then calls its own Recovery::recover() rather than the
embedded strategy.